- Convert to JS Jest Config
